+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - TaskScheduler.Builder.IntervalStep added.
 ### is.codion.swing.common.ui
 - FilterTableCellEditor.Builder added, cellEditable now configurable.
+- CalendarPanel.Builder.includeWeekNumbers added along with WEEK_NUMBERS configuration value.
 ### is.codion.framework.db
 - EntityConnection.Update.set() value parameter now @Nullable.